Move RouteFacadeContract into Contracts namespace
