+++ packages/framework/src/Facades/Route.php
@@ -3,8 +3,8 @@
 namespace Hyde\Framework\Facades;
 
 use Hyde\Framework\Contracts\PageContract;
+use Hyde\Framework\Contracts\RouteFacadeContract;
 use Hyde\Framework\Modules\Routing\Route as RouteModel;
-use Hyde\Framework\Modules\Routing\RouteFacadeContract;
 use Illuminate\Support\Collection;
 
 /**
