+++ packages/framework/src/Modules/Routing/Route.php
@@ -4,6 +4,7 @@ namespace Hyde\Framework\Modules\Routing;
 
 use Hyde\Framework\Contracts\PageContract;
 use Hyde\Framework\Contracts\RouteContract;
+use Hyde\Framework\Contracts\RouteFacadeContract;
 use Hyde\Framework\Hyde;
 use Illuminate\Support\Collection;
